Issue: If say balance after authentication enabled and customer start to enter destination before balance prompt finished it unexpectedly call wrong number. It affect customer experience who used to work with other cards on the market.
Fix: say balance function was modified. Now instead of going through if/then logic and saying balance it accumulate what to say, and then play prompt expecting input. If anything enetered, it stop saying and collect remaining digits. It returns this "early_destination", that is used in the IVR flow to suppress mention above issue.